Support native console in linux aarch64 for JRuby
I am running JRuby on a Linux aarch64.

I can see the `io/console/ffi/stty_console` is being loaded based on the warning message:
"io/console on JRuby shells out to stty for most operations"

If I manually require `io/console/ffi/native_console` it fails due to the
platform guard at the top of the file.

However, if I remove that guard it, it loads and works fine in my brief testing.

In #40, @headius confirmed there is no reason not to allow aarch64, so
adding it in this PR. Closes #40.